added TitleRepository, TitleService, TitleServiceImpl, ReportChecker, reports 1 and 4, new profile for checking the reports

+++ javaspring-team2-project-smdb/src/main/java/com/javaspring/team2/project/smdb/extraMethods/InsertMethods.java
@@ -63,7 +63,7 @@ public class InsertMethods {
         tvShow.setEndYear(Integer.parseInt((String) iterator.get("endYear")));
         tvShow.setNumberOfSeasons(Integer.parseInt((String) iterator.get("numberOfSeasons")));
         tvShow.setNumberOfEpisodes(Integer.parseInt((String) iterator.get("numberOfEpisodes")));
-        tvShow.setSmdbRating(Float.parseFloat((String) iterator.get("smdbRating")));
+        tvShow.setSmdbRating(Double.parseDouble((String) iterator.get("smdbRating")));
         tvShow.setStoryLine((String) iterator.get("storyLine"));
 
         //Adding the sets
